requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.10.0
tqdm>=4.60.0
faiss-cpu>=1.7.0
flask>=2.0.0
requests>=2.25.0
